Refactor: Relocate Sections interface to separate module.
