+++ lib/MarkdownParser.ts
@@ -2,21 +2,8 @@
 import { readFileSync } from 'node:fs'
 
 // Import internal dependencies.
+import { Section } from './types/types.js'
 import { HeadingDictionary, Dictionary } from './HeadingDictionary.js'
-
-/**
- * Represents a parsed section of a Markdown document.
- *
- * @interface Section
- * @property {number} level - The level of the heading (e.g., 1 for h1, 2 for h2).
- * @property {string} heading - The text of the heading.
- * @property {string} body - The content/body text that follows the heading.
- */
-export interface Section {
-  level: number
-  heading: string
-  body: string
-}
 
 /**
  * A class for parsing and extracting structured sections from a Markdown file.
